Deprecate rhai
