+++ auraescript/legacy/builtin/x509_certificate.rs
