+++ auraescript/legacy/examples/bin/mockd
