+++ auraescript/legacy/examples/connect.aurae
