+++ auraescript/legacy/examples/hello.aurae
