+++ auraescript/legacy/examples/status.aurae
